fix ci crash

+++ src/plugins/intel_gpu/tests/unit/test_cases/moe_scatter_reduction_gpu_test.cpp
@@ -115,21 +115,21 @@ void test_moe_scatter_reduction(bool is_caching_test, size_t k) {
             tokens_per_expert_tmp[experts_per_token[i][j]].push_back(i);
     }
 
-    std::vector<int32_t> tokens_per_expert_data;
-    std::vector<int32_t> tokens_len_per_expert_data;
-    std::vector<int32_t> experts_ids_data;
+    std::vector<size_t> tokens_per_expert_data;
+    std::vector<size_t> tokens_len_per_expert_data;
+    std::vector<size_t> experts_ids_data;
 
     for (size_t i = 0; i < tokens_per_expert_tmp.size(); ++i) {
         if (tokens_per_expert_tmp[i].empty())
             continue;
-        experts_ids_data.push_back(static_cast<int32_t>(i));
-        tokens_len_per_expert_data.push_back(static_cast<int32_t>(tokens_per_expert_tmp[i].size()));
+        experts_ids_data.push_back(i);
+        tokens_len_per_expert_data.push_back(tokens_per_expert_tmp[i].size());
         for (size_t j = 0; j < tokens_per_expert_tmp[i].size(); ++j) {
             tokens_per_expert_data.push_back(tokens_per_expert_tmp[i][j]);
         }
     }
 
-    std::vector<int32_t> expert_info_start_idx(tokens_len_per_expert_data.size(), 0);
+    std::vector<size_t> expert_info_start_idx(tokens_len_per_expert_data.size(), 0);
     for (size_t i = 1; i < tokens_len_per_expert_data.size(); ++i) {
         expert_info_start_idx[i] = expert_info_start_idx[i - 1] + tokens_len_per_expert_data[i - 1];
     }
